+++ Example/ManaKit/Utilities/MyMaintainer.swift
@@ -15,13 +15,19 @@ class MyMaintainer: Maintainer {
         startActivity(name: "updateCards")
         
         let request: NSFetchRequest<CMCard> = CMCard.fetchRequest()
-//        request.predicate = NSPredicate(format: "id != nil")
         request.sortDescriptors = [NSSortDescriptor(key: "set.releaseDate", ascending: true),
                                    NSSortDescriptor(key: "name", ascending: true)]
         let cards = try! context.fetch(request)
         var count = 0
-        print("Creating cards: \(count)/\(cards.count) \(Date())")
+        print("Updating cards: \(count)/\(cards.count) \(Date())")
         
+        let cardTypeRequest: NSFetchRequest<CMCardType> = CMCardType.fetchRequest()
+        cachedCardTypes = try! context.fetch(cardTypeRequest)
+        
+        let languageRequest: NSFetchRequest<CMLanguage> = CMLanguage.fetchRequest()
+        cachedLanguages = try! context.fetch(languageRequest)
+        let enLanguage = findLanguage(with: "en")
+
         for card in cards {
             // displayName
             var displayName: String?
@@ -47,12 +53,32 @@ class MyMaintainer: Maintainer {
                 card.myNumberOrder = order(of: collectorNumber)
             }
 
+            // myType
+            if let typeLine = card.typeLine,
+                let name = typeLine.name {
+                
+                var types = [String]()
+                for type in ManaKit.sharedInstance.typeNames {
+                    if name.contains(type) {
+                        types.append(type)
+                    }
+                }
+                
+                if types.count == 1 {
+                    card.myType = findCardType(with: types.first!, language: enLanguage!)
+                } else if types.count > 1 {
+                    card.myType = findCardType(with: "Multiple", language: enLanguage!)
+                }
+            }
+            
             // Firebase id = set.code + _ + card.name + _ + card.name+ number?
             if let _ = card.id,
                 let set = card.set,
                 let setCode = set.code,
                 let name = card.name {
                 var firebaseID = "\(setCode.uppercased())_\(name)_\(name.lowercased())"
+                    .replacingOccurrences(of: " ", with: "_")
+                    .replacingOccurrences(of: ".", with: "_")
 
                 if let variations = card.variations,
                     let array = variations.allObjects as? [CMCard] {