+++ ManaKit/Classes/CMCardFrameEffect+CoreDataClass.swift
@@ -1,0 +1,15 @@
+//
+//  CMCardFrameEffect+CoreDataClass.swift
+//  Pods
+//
+//  Created by Jovito Royeca on 07/12/2018.
+//
+//
+
+import Foundation
+import CoreData
+
+@objc(CMCardFrameEffect)
+public class CMCardFrameEffect: NSManagedObject {
+
+}