+++ ManaKit/Classes/CMCardType+CoreDataProperties.swift
@@ -18,8 +18,9 @@ extension CMCardType {
 
     @NSManaged public var name: String?
     @NSManaged public var nameSection: String?
-    @NSManaged public var cardPrintedTypes: NSSet?
-    @NSManaged public var cards: NSSet?
+    @NSManaged public var printedTypeLines: NSSet?
+    @NSManaged public var typeLines: NSSet?
+    @NSManaged public var myTypes: NSSet?
     @NSManaged public var language: CMLanguage?
 
 }
@@ -27,33 +28,50 @@ extension CMCardType {
 // MARK: Generated accessors for cardPrintedTypes
 extension CMCardType {
 
-    @objc(addCardPrintedTypesObject:)
-    @NSManaged public func addToCardPrintedTypes(_ value: CMCard)
+    @objc(addPrintedTypeLinesObject:)
+    @NSManaged public func addToPrintedTypeLines(_ value: CMCard)
 
-    @objc(removeCardPrintedTypesObject:)
-    @NSManaged public func removeFromCardPrintedTypes(_ value: CMCard)
+    @objc(removePrintedTypeLinesObject:)
+    @NSManaged public func removeFromPrintedTypeLines(_ value: CMCard)
 
-    @objc(addCardPrintedTypes:)
-    @NSManaged public func addToCardPrintedTypes(_ values: NSSet)
+    @objc(addPrintedTypeLines:)
+    @NSManaged public func addToPrintedTypeLines(_ values: NSSet)
 
-    @objc(removeCardPrintedTypes:)
-    @NSManaged public func removeFromCardPrintedTypes(_ values: NSSet)
+    @objc(removePrintedTypes:)
+    @NSManaged public func removeFromPrintedTypeLines(_ values: NSSet)
 
 }
 
-// MARK: Generated accessors for cards
+// MARK: Generated accessors for cardTypeLines
+extension CMCardType {
+    
+    @objc(addTypeLinesObject:)
+    @NSManaged public func addToTypeLines(_ value: CMCard)
+    
+    @objc(removeTypeLinesObject:)
+    @NSManaged public func removeFromTypeLines(_ value: CMCard)
+    
+    @objc(addTypeLines:)
+    @NSManaged public func addToTypeLines(_ values: NSSet)
+    
+    @objc(removeTypeLines:)
+    @NSManaged public func removeFromTypeLines(_ values: NSSet)
+    
+}
+
+// MARK: Generated accessors for cardTypeLines
 extension CMCardType {
 
-    @objc(addCardsObject:)
-    @NSManaged public func addToCards(_ value: CMCard)
+    @objc(addMyTypesObject:)
+    @NSManaged public func addToMyTypes(_ value: CMCard)
 
-    @objc(removeCardsObject:)
-    @NSManaged public func removeFromCards(_ value: CMCard)
+    @objc(removeMyTypesObject:)
+    @NSManaged public func removeFromMyTypes(_ value: CMCard)
 
-    @objc(addCards:)
-    @NSManaged public func addToCards(_ values: NSSet)
+    @objc(addMyTypes:)
+    @NSManaged public func addToMyTypes(_ values: NSSet)
 
-    @objc(removeCards:)
-    @NSManaged public func removeFromCards(_ values: NSSet)
+    @objc(removeMyTypes:)
+    @NSManaged public func removeFromMyTypes(_ values: NSSet)
 
 }